requirements.txt
langchain_community
python-dotenv
langchain_openai
unstructured[md]
chromadb